DogFighter B added

+++ Unity Project/Assets/Resources/Instances/Instance1.cs
@@ -115,7 +115,7 @@ public class Instance1 : MonoBehaviour {
 		float randomLocation = GetRandomLocation();
 
 		GameObject enemy = Resources.Load<GameObject> ("Enemies/BasicEnemy/Basic Enemy");
-		Instantiate(enemy, new Vector3(right * (float) 1.2, randomLocation, 0), Quaternion.identity);
+		Instantiate(enemy, new Vector3(right * 1.2f, randomLocation, 0), Quaternion.identity);
 
 	}
